chore(js): clean-up js e2e tests and improve speed
