+++ e2e/js/src/js-generators.ts
@@ -17,11 +17,11 @@ import {
 describe('js e2e', () => {
   let scope: string;
 
-  beforeEach(() => {
+  beforeAll(() => {
     scope = newProject();
   });
 
-  afterEach(() => cleanupProject());
+  afterAll(() => cleanupProject());
 
   it('should create libs with npm scripts', () => {
     const npmScriptsLib = uniq('npmscriptslib');
@@ -37,58 +37,6 @@ describe('js e2e', () => {
     expect(tsconfig.compilerOptions.paths).toEqual({
       [`@${scope}/${npmScriptsLib}`]: [`libs/${npmScriptsLib}/src/index.ts`],
     });
-  }, 240_000);
-
-  it('should allow wildcard ts path alias', async () => {
-    const base = uniq('base');
-    runCLI(`generate @nx/js:lib ${base} --bundler=tsc --no-interactive`);
-
-    const lib = uniq('lib');
-    runCLI(`generate @nx/js:lib ${lib} --bundler=tsc --no-interactive`);
-
-    updateFile(`libs/${base}/src/index.ts`, () => {
-      return `
-        import { ${lib} } from '@${scope}/${lib}'
-        export * from './lib/${base}';
-
-        ${lib}();
-      `;
-    });
-
-    expect(runCLI(`build ${base}`)).toContain(
-      'Done compiling TypeScript files'
-    );
-
-    updateJson('tsconfig.base.json', (json) => {
-      json['compilerOptions']['paths'][`@${scope}/${lib}/*`] = [
-        `libs/${lib}/src/*`,
-      ];
-      return json;
-    });
-
-    createFile(
-      `libs/${lib}/src/${lib}.ts`,
-      `
-export function ${lib}Wildcard() {
-  return '${lib}-wildcard';
-};
-    `
-    );
-
-    updateFile(`libs/${base}/src/index.ts`, () => {
-      return `
-        import { ${lib} } from '@${scope}/${lib}';
-        import { ${lib}Wildcard } from '@${scope}/${lib}/src/${lib}';
-        export * from './lib/${base}';
-
-        ${lib}();
-        ${lib}Wildcard();
-      `;
-    });
-
-    expect(runCLI(`build ${base}`)).toContain(
-      'Done compiling TypeScript files'
-    );
   }, 240_000);
 
   it('should create a library that can be linted and tested', async () => {
@@ -114,42 +62,6 @@ export function ${lib}Wildcard() {
       'Test Suites: 1 passed, 1 total'
     );
   }, 500_000);
-
-  it('should be able to use and be used by other libs', () => {
-    const consumerLib = uniq('consumer');
-    const producerLib = uniq('producer');
-
-    runCLI(`generate @nx/js:lib ${consumerLib} --bundler=none`);
-    runCLI(`generate @nx/js:lib ${producerLib} --bundler=none`);
-
-    updateFile(
-      `libs/${producerLib}/src/lib/${producerLib}.ts`,
-      'export const a = 0;'
-    );
-
-    updateFile(
-      `libs/${consumerLib}/src/lib/${consumerLib}.ts`,
-      `
-    import { a } from '@${scope}/${producerLib}';
-
-    export function ${consumerLib}() {
-      return a + 1;
-    }`
-    );
-    updateFile(
-      `libs/${consumerLib}/src/lib/${consumerLib}.spec.ts`,
-      `
-    import { ${consumerLib} } from './${consumerLib}';
-
-    describe('', () => {
-      it('should return 1', () => {
-        expect(${consumerLib}()).toEqual(1);
-      });
-    });`
-    );
-
-    runCLI(`test ${consumerLib}`);
-  });
 
   it('should be able to add build to non-buildable projects', () => {
     const nonBuildable = uniq('nonbuildable');
